PRSD-451: Moves/Renames some view model tests
